refactor(pages): Отрефакторил переключение между типами отображения карточек
