+++ frontend_fungi/src/app/router/index.tsx
@@ -1,11 +1,10 @@
 import { Routes, Route, Navigate } from 'react-router-dom';
 import { MainPage } from '@pages/main-page';
 import { AboutUsPage } from '@pages/about-us';
-import Publications from '@pages//publications';
 import { Encyclopedia } from '@pages/encyclopedia';
-import PublicationPage from '@pages/publications';
 import { MushroomPage } from '@pages/encyclopedia/components/MushroomPage';
-import PublicationsSec from '@pages/publications/PublicationsSec.tsx';
+import { PublicationsPage } from '@pages/publications';
+import PublicationPage from '@pages/publications/components/PublicationPage/PublicationPage';
 
 export const AppRouter = () => {
     return (
@@ -13,8 +12,7 @@ export const AppRouter = () => {
             <Route path="/" element={<Navigate to="/mainpage" />} />
             <Route path="/mainpage" element={<MainPage />} />
             <Route path="/about" element={<AboutUsPage />} />
-            <Route path="/publications" element={<Publications />} />
-            <Route path="/publications-sec" element={<PublicationsSec />} />
+            <Route path="/publications" element={<PublicationsPage />} />
             <Route path="/publications/:id" element={<PublicationPage />} />
             <Route
                 path="/encyclopedia/:mushroomId"
