+++ frontend_fungi/src/pages/Publications/components/card-publications-view/index.ts
@@ -1,0 +1,1 @@
+export { CardPublicationsView } from "./CardPublicationsView";
