+++ frontend_fungi/src/pages/Publications/components/full-size-publication-view/FullSizePublicationView.css
@@ -1,0 +1,9 @@
+.publications__container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    min-width: 1700px;
+    padding: 0;
+    margin-top: 60px;
+}
