+++ frontend_fungi/src/pages/Publications/components/full-size-publication-view/index.ts
@@ -1,0 +1,1 @@
+export { FullSizePublicationView } from "./FullSizePublicationView";
