+++ frontend_fungi/src/pages/Publications/components/publications-page/index.ts
@@ -1,0 +1,1 @@
+export { PublicationsPage } from './PublicationsPage.tsx';
