+++ frontend_fungi/src/pages/Publications/index.ts
@@ -1,0 +1,1 @@
+export { PublicationsPage } from './components/publications-page'
