+++ frontend_fungi/src/shared/ui/GridButton/index.tsx
@@ -1,15 +1,12 @@
-import { Link } from 'react-router-dom';
 
 interface IGridButton {
-    targetPath: string;
+    onClick: () => void;
 }
 
-export const GridButton: React.FC<IGridButton> = ({ targetPath }) => {
+export const GridButton: React.FC<IGridButton> = ({ onClick }) => {
     return (
-        <button className="grid-button">
-            <Link to={targetPath}>
+        <button className="grid-button" onClick={onClick}>
                 <img src="/images/svg/grid-view.svg" alt="" />
-            </Link>
         </button>
     );
 };
